Código base do novo LoadingScreen implementado
