+++ Assets/Modules/LoadingScreen.meta
@@ -1,7 +1,7 @@
 fileFormatVersion: 2
-guid: c435733b4aaace143b8667705f1fc922
+guid: ab6dee72581ee304193871b790dfdaba
 folderAsset: yes
-timeCreated: 1488488008
+timeCreated: 1502980913
 licenseType: Free
 DefaultImporter:
   userData:
