+++ Assets/Modules/LoadingScreen/Prefabs.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 92648d48ce4ddf746a17be369284c6a4
-timeCreated: 1491847463
+guid: d50c5936d2725a64f96028deef7760d8
+folderAsset: yes
+timeCreated: 1502980925
 licenseType: Free
 DefaultImporter:
   userData:
