+++ Assets/Modules/LoadingScreen/Prefabs/LoadingScreen.prefab.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 48088b230f9db3f4290f5b1a859b1a64
+timeCreated: 1502981120
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
