--- Assets/Modules/LoadingScreen/Scenes.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 15745a06f21596049808c22f2cc4854c
-folderAsset: yes
-timeCreated: 1499830770
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
