--- Assets/Modules/LoadingScreen/Scenes/LoadingScreen.unity.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0f8d99f81d331674da32d211337bbff5
-timeCreated: 1488685818
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
