+++ Assets/Modules/LoadingScreen/Scripts.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 3a6d47a829317c14f83f043402b06964
+guid: 4a382746d5913ca429ef98ae4ea6a7d3
 folderAsset: yes
-timeCreated: 1495996878
-licenseType: Pro
+timeCreated: 1502980919
+licenseType: Free
 DefaultImporter:
   userData: 
   assetBundleName:
