+++ Assets/Modules/LoadingScreen/Scripts/LoadingScreen.cs
@@ -1,0 +1,45 @@
+﻿using System.Collections;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+public class LoadingScreen : MonoBehaviour
+{
+    public static LoadingScreen Instance;
+
+    // Use this for initialization
+    private void Awake()
+    {
+        if (Instance == null)
+        {
+            Instance = this;
+        }
+        else if (Instance != this)
+        {
+            Destroy(gameObject);
+            return;
+        }
+
+        DontDestroyOnLoad(gameObject);
+    }
+
+    private IEnumerator LoadSceneAsync(int sceneIndex)
+    {
+        var operation = SceneManager.LoadSceneAsync(sceneIndex);
+        while (!operation.isDone)
+        {
+            float progress = Mathf.Clamp01(operation.progress / 0.9f);
+
+            Debug.LogFormat("LoadingScene - sceneIndex:{0} progress:{1}",
+                sceneIndex, progress);
+
+            //Slider.value = progress;
+
+            yield return null;
+        }
+    }
+
+    public void LoadScene(int sceneIndex)
+    {
+        StartCoroutine(LoadSceneAsync(sceneIndex));
+    }
+}
