+++ Assets/Modules/LoadingScreen/Scripts/LoadingScreen.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: d030a77e0567c1042956531fc89a28d6
-timeCreated: 1499829389
+guid: 2b13bb01c015e9c45a60dccd62d55ceb
+timeCreated: 1502981102
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
