--- Assets/Modules/LoadingScreen/Scripts/LoadingScreenManager.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: d4e98a337d4229440bd70d727a46c42d
-timeCreated: 1488486370
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
