--- Assets/Scripts/Debug/DebugToggle.cs
@@ -1,9 +1,0 @@
-﻿using UnityEngine;
-
-public class DebugToggle : MonoBehaviour
-{
-	private void Start()
-	{
-        this.gameObject.SetActive(GameDataManager.Instance.IsDebug);
-    }
-}
