--- Assets/Scripts/Debug/DebugToggle.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 36a531f62f7dab543832a5763d7e7a44
-timeCreated: 1488656130
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
