--- Assets/Scripts/GameData/Data.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 2c9510b3fcc0d4f4fbede1cad495721f
-folderAsset: yes
-timeCreated: 1496013325
-licenseType: Pro
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
