--- Assets/Scripts/GameData/GameData.cs
@@ -1,5 +1,0 @@
-﻿[System.Serializable]
-public class GameData
-{
-    public bool DebugMode;
-}
