--- Assets/Scripts/GameData/GameData.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: ebdc405392238904092586aa8da05bac
-timeCreated: 1487710872
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
