--- Assets/Scripts/GameData/GameDataManager.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 79f8ceba70c05e74e8257047156a705e
-timeCreated: 1487718460
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
