+++ Assets/Scripts/GameData/Pacient.cs
