--- Assets/Scripts/Startup.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 27dab380519ecd841822d933c9d6c332
-timeCreated: 1489422150
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
